requirements.txt
chainlit>=2.6.0,<3.0.0
openai>=1.40.0
anthropic>=0.34.0
groq>=0.9.0
google-genai>=0.3.0
python-dotenv>=1.0.1
langchain-core>=0.2.0